frontend/requirements.txt
# Plex Watcher Frontend Requirements
# Install with: pip install -r requirements.txt

# Web Framework
streamlit>=1.50.0

# Async HTTP Client
httpx>=0.28.1

# Environment Variables
python-dotenv>=1.1.1

# Data Processing
pandas>=2.3.3
